Add build config for react refactor
