+++ hermes-extension/hermes-react-refactor/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "hermes-react-refactor",
+  "version": "1.0.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "build": "tsc"
+  },
+  "devDependencies": {
+    "typescript": "^5.8.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "@types/react": "^18.2.7",
+    "@types/react-dom": "^18.2.4",
+    "@types/chrome": "^0.0.234",
+    "@types/node": "^20.8.10",
+    "@reduxjs/toolkit": "^1.9.7",
+    "react-redux": "^8.1.1",
+    "@types/react-redux": "^7.1.25"
+  }
+}
