+++ hermes-extension/hermes-react-refactor/src/App.tsx
@@ -4,10 +4,10 @@ import React, { useState, useRef, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { RootState, AppDispatch } from './store';
 import { SettingsPanel } from './components/SettingsPanel';
-import MacroPanel from './components/MacroPanel';
-import { TrainerPanel } from './components/TrainerPanel';
+import MacroPanel from './components/macroPanel';
+import { TrainerPanel } from './components/trainerPanel';
 import { useDraggable } from './hooks/useDraggable';
-import { macroEngine } from './services/macroEngineService';
+import { macroEngine } from './content';
 import { fillForm } from './services/formFillerService';
 import { startHeuristicTraining } from './services/trainerService';
 import { SkippedField } from './services/heuristicsService';
