+++ hermes-extension/hermes-react-refactor/src/three.d.ts
@@ -1,0 +1,1 @@
+declare module 'https://cdn.jsdelivr.net/npm/three@0.161.0/build/three.module.js';
