+++ hermes-extension/hermes-react-refactor/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "target": "ES2020",
+    "module": "esnext",
+    "jsx": "react-jsx",
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "outDir": "dist",
+    "types": ["chrome", "node"]
+  },
+  "include": ["src/**/*", "src/three.d.ts"]
+}
