+++ hermes-extension/tsconfig.json
@@ -8,7 +8,8 @@
     "allowImportingTsExtensions": true,
     "jsx": "react-jsx",
     "moduleResolution": "node",
-    "esModuleInterop": true
+    "esModuleInterop": true,
+    "types": ["chrome", "jest", "node"]
   },
   "include": ["src/**/*"]
 }
